+++ src/main/java/com/albion/common/graph/algorithms/DijkstraV1.java
@@ -1,4 +1,4 @@
-package com.albion.common.search.graph;
+package com.albion.common.graph.algorithms;
 
 import com.albion.common.graph.core.Directions;
 import com.albion.common.graph.core.v2.Edge;
@@ -14,19 +14,19 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-public class DijkstraV2 extends BaseDijkstra<String> {
+public class DijkstraV1 extends BaseDijkstra<Integer>{
 
-    public DijkstraV2(Graph<String> g) {
+    public DijkstraV1(Graph<Integer> g) {
         super(g);
     }
 
     public void parseInput(String filePath) {
-        HashMap<String, Vertex<String>> map = parseInputIDAsString(filePath);
+        HashMap<Integer, Vertex<Integer>> map = parseInputIDAsInteger(filePath);
         graph.setVerticesMap(map);
     }
 
-    public HashMap<String, Vertex<String>> parseInputIDAsString(String filePath){
-        HashMap<String, Vertex<String>> map = new HashMap<>();
+    public HashMap<Integer, Vertex<Integer>> parseInputIDAsInteger(String filePath){
+        HashMap<Integer, Vertex<Integer>> map = new HashMap<>();
         File inputFile = new File(filePath);
 
         try {
@@ -37,8 +37,8 @@ public class DijkstraV2 extends BaseDijkstra<String> {
                 Node nNode = vertexList.item(i);
                 if (nNode.getNodeType() == Node.ELEMENT_NODE) {
                     Element elem = (Element) nNode;
-                    String vertexId = elem.getAttribute("id");
-                    Vertex<String> vertex = new Vertex<>(vertexId);
+                    int vertexId = Integer.parseInt(elem.getAttribute("id"));
+                    Vertex<Integer> vertex = new Vertex<>(vertexId);
                     NodeList edgeList = elem.getElementsByTagName("edge");
                     addEdgesToAVertex(vertex, edgeList);
                     map.put(vertexId, vertex);
@@ -51,8 +51,8 @@ public class DijkstraV2 extends BaseDijkstra<String> {
         return map;
     }
 
-    public Vertex<String> addEdgesToAVertex(Vertex<String> vertex, NodeList list){
-        List<Edge<String>> edgeList = new ArrayList<>();
+    private Vertex<Integer> addEdgesToAVertex(Vertex<Integer> vertex, NodeList list){
+        List<Edge<Integer>> edgeList = new ArrayList<>();
         for(int j = 0; j < list.getLength(); j++){
             Node mNode = list.item(j);
             if (mNode.getNodeType() == Node.ELEMENT_NODE) {
@@ -75,7 +75,7 @@ public class DijkstraV2 extends BaseDijkstra<String> {
                 if("".equals(weight)){
                     weight = "0";
                 }
-                Edge<String> edge = new Edge<>(vertex.getId(), id, way, Integer.parseInt(weight));
+                Edge<Integer> edge = new Edge<>(vertex.getId(), Integer.parseInt(id), way, Integer.parseInt(weight));
                 edgeList.add(edge);
             }
         }
