+++ src/main/java/com/albion/common/tree/bst/traversal/Traversal.java
@@ -1,4 +1,4 @@
-package com.albion.common.search.bst;
+package com.albion.common.tree.bst.traversal;
 
 public interface Traversal {
 	void traverse();
