+++ src/test/java/com/albion/common/graph/algorithms/DijkstraV1Test.java
@@ -1,5 +1,6 @@
-package com.albion.common.search.graph;
+package com.albion.common.graph.algorithms;
 
+import com.albion.common.graph.algorithms.DijkstraV1;
 import com.albion.common.graph.core.v2.Graph;
 import com.albion.common.graph.core.v2.Vertex;
 import org.testng.Assert;
@@ -8,36 +9,36 @@ import org.testng.annotations.Test;
 
 import java.util.List;
 
-public class DijkstraV2Test {
+public class DijkstraV1Test {
     @DataProvider(name = "dp01")
     public Object[][] makeData() {
         return new Object[][] {
-                {"src/test/resources/graph03.xml", "San Francisco", "Boston", 14},
-                {"src/test/resources/graph04.xml", "San Francisco", "Chicago", 13}
+                {"src/test/resources/graph01.xml", 1, 4, 14},
+                {"src/test/resources/graph02.xml", 1, 5, 13}
         };
     }
 
     @Test(dataProvider = "dp01")
-    public void testFindShortestPath(String filePath, String src, String destination, int expectedCost) throws Exception {
-        Graph<String> g = new Graph<>();
-        DijkstraV2 dijkstra = new DijkstraV2(g);
+    public void testFindShortestPath(String filePath, int src, int destination, int expectedCost) throws Exception {
+        Graph<Integer> g = new Graph<>();
+        DijkstraV1 dijkstra = new DijkstraV1(g);
         dijkstra.parseInput(filePath);
 
-        List<Vertex<String>> result = dijkstra.findShortestDistance(src, destination);
+        List<Vertex<Integer>> result = dijkstra.findShortestDistance(src, destination);
         System.out.println("============");
         System.out.println(g.toString());
 
         System.out.println("[Shortest Path]");
         int size = result.size();
         for (int i = 0; i < size; i++) {
-            Vertex v = result.get(i);
+            Vertex<Integer> v = result.get(i);
             System.out.print("[ID:" + v.getId() + " cost:" + v.getCost() + "]");
             if(i != size - 1) {
                 System.out.print("-->");
             }
         }
         System.out.println("");
-        Vertex target = result.get(result.size() - 1);
+        Vertex<Integer> target = result.get(result.size() - 1);
         Assert.assertEquals(target.getCost(), expectedCost);
     }
 }
