+++ src/test/java/com/albion/common/tree/bst/traversal/InOrderTraversalTest.java
@@ -1,16 +1,17 @@
-package com.albion.common.search.bst;
+package com.albion.common.tree.bst.traversal;
 
 import com.albion.common.tree.bst.BinarySearchTree;
+import com.albion.common.tree.bst.traversal.InOrderTraversal;
+import com.albion.common.tree.bst.traversal.Traversal;
 import com.albion.common.tree.utils.BinaryTreePrinter;
 import com.albion.common.tree.node.BinarySearchTreeNode;
 import org.testng.annotations.Test;
 
-
-public class BinaryTreeInversionTest {
+public class InOrderTraversalTest {
     @Test
-    public void testInvert() throws Exception {
+    public void testTraverse() throws Exception {
         Integer[] array = {15, 6, 18, 3, 7, 17, 20, 2, 4, 13, 9};
-        BinarySearchTree<Integer> bt = new BinarySearchTree<>();
+        BinarySearchTree<Integer> bt = new BinarySearchTree();
 
         for(Integer x : array) {
             bt.insert(x);
@@ -19,9 +20,8 @@ public class BinaryTreeInversionTest {
         BinarySearchTreeNode<Integer> root = bt.getRoot();
         BinaryTreePrinter.printNode(root);
 
-        BinaryTreeInversion<Integer> bti = new BinaryTreeInversion<>(root);
-        bti.invert();
-        BinaryTreePrinter.printNode(root);
+        Traversal traversal = new InOrderTraversal<>(root);
+        traversal.traverse();
     }
 
 }
