+++ src/test/resources/graph01.xml
@@ -1,23 +1,23 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <verticesMap>
-    <vertices>
-	    <vertex id="1">
-	        <edge id="2" weight="8" direction="BOTH" />
-	        <edge id="3" weight="9" direction="BOTH" /> 
-	    </vertex>
-	    <vertex id="2" >
-	        <edge id="1" weight="8" direction="BOTH" />
-	        <edge id="3" weight="6" direction="BOTH" /> 
-	   	    <edge id="4" weight="7" direction="BOTH" />      
-	    </vertex>
-	    <vertex id="3" >
-	        <edge id="1" weight="9" direction="BOTH" />
-	        <edge id="2" weight="6" direction="BOTH" /> 
-	   	    <edge id="4" weight="5" direction="BOTH" />      
-	    </vertex>
-	    <vertex id="4" >
-	        <edge id="2" weight="7" direction="BOTH" />
-	        <edge id="3" weight="5" direction="BOTH" />  
-	    </vertex>		    	    
-    </vertices>
+	<vertices>
+		<vertex id="1">
+			<edge id="2" weight="8" direction="BOTH" />
+			<edge id="3" weight="9" direction="BOTH" />
+		</vertex>
+		<vertex id="2" >
+			<edge id="1" weight="8" direction="BOTH" />
+			<edge id="3" weight="6" direction="BOTH" />
+			<edge id="4" weight="7" direction="BOTH" />
+		</vertex>
+		<vertex id="3" >
+			<edge id="1" weight="9" direction="BOTH" />
+			<edge id="2" weight="6" direction="BOTH" />
+			<edge id="4" weight="5" direction="BOTH" />
+		</vertex>
+		<vertex id="4" >
+			<edge id="2" weight="7" direction="BOTH" />
+			<edge id="3" weight="5" direction="BOTH" />
+		</vertex>
+	</vertices>
 </verticesMap>
